backend/requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
tensorflow>=2.13.0
opencv-python>=4.8.0
numpy>=1.24.0
Werkzeug>=2.3.0
Pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
